♻️ refactor: move AgentRuntime
