+++ src/app/(backend)/api/chat/[provider]/route.ts
@@ -1,11 +1,11 @@
 import { createErrorResponse } from '@/app/(backend)/api/errorResponse';
 import { AgentRuntime, ChatCompletionErrorPayload } from '@/libs/agent-runtime';
+import { createTraceOptions, initAgentRuntimeWithUserPayload } from '@/server/modules/AgentRuntime';
 import { ChatErrorType } from '@/types/fetch';
 import { ChatStreamPayload } from '@/types/openai/chat';
 import { getTracePayload } from '@/utils/trace';
 
 import { checkAuth } from '../../middleware/auth';
-import { createTraceOptions, initAgentRuntimeWithUserPayload } from '../agentRuntime';
 
 export const runtime = 'edge';
